requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
pydicom>=2.4.0          # For reading DICOM (.dcm) medical image files
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0
pytorch-fid>=0.3.0      # For FID evaluation metric (optional)
kaggle>=1.5.16          # For downloading datasets via Kaggle API (optional)
